+++ src/ladb_opencutlist/ruby/worker/common/common_drawing_decomposition_worker.rb
@@ -30,27 +30,47 @@ module Ladb::OpenCutList
     EDGE_VALIDATOR_STRAY = 2
     EDGE_VALIDATOR_STRAY_COPLANAR = 3
 
-    def initialize(path, settings = {})
+    def initialize(path,
+
+                   input_local_x_axis: X_AXIS,
+                   input_local_y_axis: Y_AXIS,
+                   input_local_z_axis: Z_AXIS,
+
+                   input_plane_manipulator: nil,
+                   input_line_manipulator: nil,
+
+                   origin_position: ORIGIN_POSITION_DEFAULT,
+
+                   ignore_faces: false,
+                   ignore_surfaces: false,
+                   face_validator: FACE_VALIDATOR_ALL,
+                   face_recursive: true,
+
+                   ignore_edges: false,
+                   edge_validator: EDGE_VALIDATOR_ALL,
+                   edge_recursive: true
+
+    )
 
       @path = path
 
-      @input_local_x_axis = settings.fetch('input_local_x_axis', X_AXIS)
-      @input_local_y_axis = settings.fetch('input_local_y_axis', Y_AXIS)
-      @input_local_z_axis = settings.fetch('input_local_z_axis', Z_AXIS)
+      @input_local_x_axis = input_local_x_axis
+      @input_local_y_axis = input_local_y_axis
+      @input_local_z_axis = input_local_z_axis
 
-      @input_plane_manipulator = settings.fetch('input_plane_manipulator', nil)
-      @input_line_manipulator = settings.fetch('input_line_manipulator', nil)
+      @input_plane_manipulator = input_plane_manipulator
+      @input_line_manipulator = input_line_manipulator
 
-      @origin_position = settings.fetch('origin_position', ORIGIN_POSITION_DEFAULT)
+      @origin_position = origin_position
 
-      @ignore_faces = settings.fetch('ignore_faces', false)
-      @ignore_surfaces = settings.fetch('ignore_surfaces', false)
-      @face_validator = settings.fetch('face_validator', FACE_VALIDATOR_ALL)
-      @face_recursive = settings.fetch('face_recursive', true)
+      @ignore_faces = ignore_faces
+      @ignore_surfaces = ignore_surfaces
+      @face_validator = face_validator
+      @face_recursive = face_recursive
 
-      @ignore_edges = settings.fetch('ignore_edges', false)
-      @edge_validator = settings.fetch('edge_validator', EDGE_VALIDATOR_ALL)
-      @edge_recursive = settings.fetch('edge_recursive', true)
+      @ignore_edges = ignore_edges
+      @edge_validator = edge_validator
+      @edge_recursive = edge_recursive
 
     end
 
@@ -100,29 +120,32 @@ module Ladb::OpenCutList
       origin = ORIGIN.transform(transformation)
       if @input_plane_manipulator
 
-        drawing_def.picked_plane_manipulator = @input_plane_manipulator
-        drawing_def.picked_line_manipulator = @input_line_manipulator
+        drawing_def.input_plane_manipulator = @input_plane_manipulator
+        drawing_def.input_line_manipulator = @input_line_manipulator
 
         if @input_line_manipulator
 
-          x_axis, y_axis, z_axis, drawing_def.picked_line_manipulator = _get_input_axes(drawing_def.picked_plane_manipulator, drawing_def.picked_line_manipulator)
+          x_axis, y_axis, z_axis, drawing_def.input_line_manipulator = _get_input_axes(drawing_def.input_plane_manipulator, drawing_def.input_line_manipulator)
 
         else
 
-          input_inner_plane_manipulator = PlaneManipulator.new(@input_plane_manipulator.plane, @input_plane_manipulator.transformation * transformation.inverse)
+          input_inner_plane_manipulator = PlaneManipulator.new(@input_plane_manipulator.face.plane, @input_plane_manipulator.transformation * transformation.inverse)
 
           if input_inner_plane_manipulator.normal.parallel?(@input_local_z_axis) || input_inner_plane_manipulator.normal.parallel?(@input_local_y_axis)
-            z_axis = drawing_def.picked_plane_manipulator.normal
+            z_axis = drawing_def.input_plane_manipulator.normal
             x_axis = @input_local_x_axis.transform(transformation)
             x_axis.reverse! if TransformationUtils.flipped?(transformation)
             y_axis = z_axis * x_axis
           elsif input_inner_plane_manipulator.normal.parallel?(@input_local_x_axis)
-            z_axis = drawing_def.picked_plane_manipulator.normal
+            z_axis = drawing_def.input_plane_manipulator.normal
             x_axis = @input_local_y_axis.transform(transformation)
             x_axis.reverse! if TransformationUtils.flipped?(transformation)
             y_axis = z_axis * x_axis
           else
-            x_axis, y_axis, z_axis, drawing_def.picked_line_manipulator = _get_input_axes(drawing_def.picked_plane_manipulator, nil)
+
+            puts 'oula'
+
+            x_axis, y_axis, z_axis, drawing_def.input_line_manipulator = _get_input_axes(drawing_def.input_plane_manipulator, nil)
           end
 
         end
@@ -147,8 +170,8 @@ module Ladb::OpenCutList
       ttai = tai * transformation
 
       drawing_def.transformation = ta
-      drawing_def.picked_plane_manipulator.transformation = tai * drawing_def.picked_plane_manipulator.transformation unless drawing_def.picked_plane_manipulator.nil?
-      drawing_def.picked_line_manipulator.transformation = tai * drawing_def.picked_line_manipulator.transformation unless drawing_def.picked_line_manipulator.nil?
+      drawing_def.input_plane_manipulator.transformation = tai * drawing_def.input_plane_manipulator.transformation unless drawing_def.input_plane_manipulator.nil?
+      drawing_def.input_line_manipulator.transformation = tai * drawing_def.input_line_manipulator.transformation unless drawing_def.input_line_manipulator.nil?
 
       # STEP 2 : Populate faces and edges manipulators
 
@@ -156,23 +179,23 @@ module Ladb::OpenCutList
       unless @ignore_faces
 
         validator = nil
-        if drawing_def.picked_plane_manipulator
+        if drawing_def.input_plane_manipulator
           case @face_validator
           when FACE_VALIDATOR_ONE
             validator = lambda { |face_manipulator|
-              face_manipulator == drawing_def.picked_plane_manipulator
+              face_manipulator == drawing_def.input_plane_manipulator
             }
           when FACE_VALIDATOR_COPLANAR
             validator = lambda { |face_manipulator|
-              face_manipulator.coplanar?(drawing_def.picked_plane_manipulator)
+              face_manipulator.coplanar?(drawing_def.input_plane_manipulator)
             }
           when FACE_VALIDATOR_PARALLEL
             validator = lambda { |face_manipulator|
-              face_manipulator.parallel?(drawing_def.picked_plane_manipulator)
+              face_manipulator.parallel?(drawing_def.input_plane_manipulator)
             }
           when FACE_VALIDATOR_EXPOSED
             validator = lambda { |face_manipulator|
-              !face_manipulator.perpendicular?(drawing_def.picked_plane_manipulator) && drawing_def.picked_plane_manipulator.angle_between(face_manipulator) < Math::PI / 2.0
+              !face_manipulator.perpendicular?(drawing_def.input_plane_manipulator) && drawing_def.input_plane_manipulator.angle_between(face_manipulator) < Math::PI / 2.0
             }
           end
         end
@@ -189,8 +212,8 @@ module Ladb::OpenCutList
         case @edge_validator
         when EDGE_VALIDATOR_COPLANAR
           validator = lambda { |edge_manipulator|
-            return false if drawing_def.picked_plane_manipulator.nil?
-            edge_manipulator.direction.perpendicular?(drawing_def.picked_plane_manipulator.normal) && edge_manipulator.position.on_plane?(drawing_def.picked_plane_manipulator.plane)
+            return false if drawing_def.input_plane_manipulator.nil?
+            edge_manipulator.direction.perpendicular?(drawing_def.input_plane_manipulator.normal) && edge_manipulator.position.on_plane?(drawing_def.input_plane_manipulator.plane)
           }
         when EDGE_VALIDATOR_STRAY
           validator = lambda { |edge_manipulator|
@@ -198,9 +221,9 @@ module Ladb::OpenCutList
           }
         when EDGE_VALIDATOR_STRAY_COPLANAR
           validator = lambda { |edge_manipulator|
-            return false if drawing_def.picked_plane_manipulator.nil?
+            return false if drawing_def.input_plane_manipulator.nil?
             if edge_manipulator.edge.faces.empty?
-              edge_manipulator.direction.perpendicular?(drawing_def.picked_plane_manipulator.normal) && edge_manipulator.position.on_plane?(drawing_def.picked_plane_manipulator.plane)
+              edge_manipulator.direction.perpendicular?(drawing_def.input_plane_manipulator.normal) && edge_manipulator.position.on_plane?(drawing_def.input_plane_manipulator.plane)
             else
               false
             end
@@ -250,18 +273,18 @@ module Ladb::OpenCutList
 
     private
 
-    def _get_input_axes(picked_plane_manipulator, picked_line_manipulator = nil)
+    def _get_input_axes(input_plane_manipulator, input_line_manipulator = nil)
 
-      if picked_line_manipulator.nil? || !picked_plane_manipulator.normal.perpendicular?(picked_line_manipulator.direction)
-        picked_line_manipulator = EdgeManipulator.new(picked_plane_manipulator.longest_outer_edge, picked_plane_manipulator.transformation)
+      if input_line_manipulator.nil? || !input_plane_manipulator.normal.perpendicular?(input_line_manipulator.direction)
+        input_line_manipulator = EdgeManipulator.new(input_plane_manipulator.longest_outer_edge, input_plane_manipulator.transformation)
       end
 
-      z_axis = picked_plane_manipulator.normal
-      x_axis = picked_line_manipulator.direction
-      x_axis.reverse! if picked_line_manipulator.respond_to?(:reversed_in?) && picked_plane_manipulator.respond_to?(:face) && picked_line_manipulator.reversed_in?(picked_plane_manipulator.face)
+      z_axis = input_plane_manipulator.normal
+      x_axis = input_line_manipulator.direction
+      x_axis.reverse! if input_line_manipulator.respond_to?(:reversed_in?) && input_plane_manipulator.respond_to?(:face) && input_line_manipulator.reversed_in?(input_plane_manipulator.face)
       y_axis = z_axis.cross(x_axis).normalize
 
-      [ x_axis, y_axis, z_axis, picked_line_manipulator ]
+      [ x_axis, y_axis, z_axis, input_line_manipulator ]
     end
 
     def _populate_face_manipulators(drawing_def, entities, transformation = IDENTITY, recursive = true, &validator)