+++ CHANGELOG.md
@@ -8,6 +8,7 @@ in 6.x, 5.x, 4.x, 3.x, 2.x, 1.x and 0.x versions.
 
     * Added Move Axes action on Smart Axes Tool
     * Added Export Paths action on Smart Export Tool
+    * Added Switch YZ axes option to Smart Export Tool on Export 3D part action
     * Improved smart tools picker
     * Improved surface detection